+++ noodles-bam/examples/bam_idxstats.rs
@@ -9,6 +9,7 @@
 use std::{env, fs::File, path::PathBuf};
 
 use noodles_bam::{self as bam, bai};
+use noodles_csi::BinningIndexReferenceSequence;
 use noodles_sam as sam;
 
 fn main() -> Result<(), Box<dyn std::error::Error>> {